backend/src/requirements.txt
bottle
psycopg[binary]
google-auth
requests
pytesseract
openai
